Add in some tips and workarounds for problems
I've added in some tips and workarounds from problems I've run into from hosting an S1X server myself, along with system requirements I've seen in the real world. I've also added in a link to the S1X logger script as this is needed for admin tools to work.